some refactoring. Moved lots of functions to a new file. Included test shape files
